Updated all modules

+++ dao/src/main/java/by/freebook/dao/repository/Repository.java
@@ -6,11 +6,11 @@ import java.util.Optional;
 public interface Repository<T> {
     List<T> findAll();
 
-    Optional<T> find(Integer id);
+    Optional<T> find(Long id);
 
     Optional<T> save(T entity);
 
     Optional<T> update(T entity);
 
-    Optional<T> remove(Integer id);
+    Optional<T> remove(Long id);
 }
